Add unit test for upload show and instruction
